Move up command into its own module
